@@ -3,127 +3,94 @@
   "version": "0.0.1",
   "description": "Complete react-redux-firebase Material-UI app.",
   "main": "index.js",
-  "engines": {
-    "node": ">=4.2.0",
-    "npm": "^3.0.0"
-  },
   "scripts": {
     "clean": "rimraf dist",
+    "compile": "node build/scripts/compile",
+    "build": "npm run clean && cross-env NODE_ENV=production npm run compile",
+    "build:prod": "echo \"build:prod is deprecated. Use npm run build.\" && npm run build",
+    "start": "cross-env NODE_ENV=development webpack-dashboard -p 1234 -c cyan -m -t material -- node build/scripts/start",
+    "start:simple": "cross-env NODE_ENV=development node build/scripts/start",
+    "lint": "eslint .",
+    "lint:fix": "npm run format && npm run lint -- --fix",
     "format": "prettier --single-quote --no-semi --trailing-comma none --write \"src/**/*.js\"",
-    "compile": "better-npm-run compile",
-    "lint": "eslint src tests bin build server",
-    "lint:fix": "npm run lint -- --fix",
-    "start": "better-npm-run start",
-    "dev": "better-npm-run dev",
-    "build": "better-npm-run build",
-    "build:dev": "better-npm-run build:dev",
-    "build:prod": "better-npm-run build:prod"
-  },
-  "betterScripts": {
-    "compile": {
-      "command": "node bin/compile",
-      "env": {
-        "DEBUG": "app:*"
-      }
-    },
-    "build": {
-      "command": "npm run clean && npm run compile",
-      "env": {
-        "DEBUG": "app:*"
-      }
-    },
-    "build:dev": {
-      "command": "npm run build",
-      "env": {
-        "NODE_ENV": "development",
-        "DEBUG": "app:*"
-      }
-    },
-    "build:prod": {
-      "command": "npm run build",
-      "env": {
-        "NODE_ENV": "production",
-        "DEBUG": "app:*"
-      }
-    },
-    "start": {
-      "command": "node bin/dev-server",
-      "env": {
-        "DEBUG": "app:*"
-      }
-    }
+    "prepush": "npm run lint"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/prescottprue/material.git"
+    "url": "git+https://github.com/testuser/material.git"
   },
-  "author": "prescottprue (https://github.com/prescottprue)",
+  "author": "testuser (https://github.com/testuser)",
   "license": "MIT",
   "dependencies": {
-    "lodash": "^4.17.2",
-    "material-ui": "0.16.0",
-    "normalize.css": "^4.1.1",
-    "react": "15.3.2",
-    "react-dom": "15.3.2",
-    "react-google-button": "^0.1.0",
-    "react-redux": "^4.4.5",
-    "react-redux-firebase": "^1.4.0-rc.1",
-    "react-router": "^2.8.0",
-    "react-tap-event-plugin": "1.0.0",
+    "lodash": "^4.17.4",
+    "material-ui": "^0.19.1",
+    "normalize.css": "^7.0.0",
+    "object-assign": "^4.1.1",
+    "promise": "^7.1.1",
+    "prop-types": "^15.5.10",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-google-button": "^0.3.0",
+    "react-redux": "^5.0.4",
+    "react-redux-firebase": "next",
+    "react-router": "^3.0.0",
+    "react-tap-event-plugin": "^2.0.1",
     "redux": "^3.6.0",
     "redux-auth-wrapper": "^1.0.0",
     "redux-form": "^6.6.1",
-    "redux-thunk": "^2.2.0"
+    "redux-form-material-ui": "^4.2.0",
+    "redux-thunk": "^2.2.0",
+    "redbox-react": "^1.3.6",
+    "whatwg-fetch": "^2.0.3"
   },
   "devDependencies": {
-    "babel-core": "^6.22.1",
-    "babel-eslint": "^6.0.4",
-    "babel-loader": "^6.2.5",
-    "babel-plugin-istanbul": "^2.0.1",
-    "babel-plugin-lodash": "^3.2.10",
+    "babel-core": "^6.24.1",
+    "babel-eslint": "^7.2.3",
+    "babel-loader": "^7.0.0",
+    "babel-plugin-lodash": "^3.2.11",
+    "babel-plugin-syntax-dynamic-import": "^6.18.0",
+    "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
+    "babel-plugin-transform-object-rest-spread": "^6.23.0",
     "babel-plugin-transform-runtime": "^6.15.0",
-    "babel-polyfill": "^6.16.0",
-    "babel-preset-es2015": "^6.14.0",
-    "babel-preset-react": "^6.11.1",
-    "babel-preset-stage-0": "^6.3.13",
-    "babel-runtime": "^6.11.6",
-    "better-npm-run": "^0.0.13",
+    "babel-preset-env": "^1.4.0",
+    "babel-preset-react": "^6.24.1",
+    "babel-runtime": "^6.20.0",
+    "chalk": "^1.1.3",
+    "compression": "^1.6.2",
     "connect-history-api-fallback": "^1.3.0",
-    "css-loader": "^0.25.0",
-    "cssnano": "^3.7.4",
-    "debug": "^2.2.0",
-    "enzyme": "^2.5.1",
-    "eslint": "^3.9.1",
+    "cross-env": "^5.0.0",
+    "css-loader": "^0.28.1",
+    "eslint": "^3.19.0",
     "eslint-config-prettier": "^2.3.0",
-    "eslint-config-standard": "^6.2.1",
-    "eslint-config-standard-react": "^4.2.0",
-    "eslint-plugin-babel": "^3.3.0",
+    "eslint-config-standard": "^10.2.1",
+    "eslint-config-standard-react": "^5.0.0",
+    "eslint-plugin-babel": "^4.1.1",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-node": "^4.2.2",
+    "eslint-plugin-promise": "^3.5.0",
     "eslint-plugin-prettier": "^2.1.2",
-    "eslint-plugin-promise": "^3.3.1",
-    "eslint-plugin-react": "^6.0.0",
-    "eslint-plugin-standard": "^2.0.0",
+    "eslint-plugin-react": "^7.0.1",
+    "eslint-plugin-standard": "^3.0.1",
     "express": "^4.14.0",
-    "extract-text-webpack-plugin": "^1.0.0",
-    "file-loader": "^0.9.0",
-    "fs-extra": "^1.0.0",
-    "html-webpack-plugin": "^2.22.0",
-    "imports-loader": "^0.6.5",
-    "ip": "^1.1.2",
-    "json-loader": "^0.5.4",
-    "mocha": "^3.0.1",
-    "node-sass": "^3.7.0",
-    "nodemon": "^1.10.2",
-    "postcss-loader": "^0.13.0",
-    "prettier": "^1.5.2",
-    "redbox-react": "^1.2.10",
-    "rimraf": "^2.5.4",
-    "sass-loader": "^4.0.0",
-    "style-loader": "^0.13.1",
-    "url-loader": "^0.5.6",
-    "webpack": "^1.12.14",
-    "webpack-dev-middleware": "1.10.1",
-    "webpack-hot-middleware": "^2.12.2",
-    "yargs": "^6.3.0"
+    "extract-text-webpack-plugin": "^2.1.0",
+    "favicons-webpack-plugin": "0.0.7",
+    "figures": "^2.0.0",
+    "file-loader": "^0.11.1",
+    "fs-extra": "^3.0.1",
+    "html-webpack-plugin": "^2.29.0",
+    "husky": "^0.14.3",
+    "ip": "^1.1.5",
+    "node-sass": "^4.5.3",
+    "prettier": "^1.5.3",
+    "rimraf": "^2.6.1",
+    "sass-loader": "^6.0.5",
+    "style-loader": "^0.17.0",
+    "url-loader": "^0.5.8",
+    "webpack": "^2.5.1",
+    "webpack-dashboard": "^0.4.0",
+    "webpack-dev-middleware": "^1.11.0",
+    "webpack-hot-middleware": "^2.13.2",
+    "yargs": "^8.0.1"
   }
 }
