@@ -1,18 +1,18 @@
 import React from 'react'
-import Navbar from 'containers/Navbar/Navbar'
+import PropTypes from 'prop-types'
+import Navbar from 'containers/Navbar'
 import classes from './CoreLayout.scss'
 import 'styles/core.scss'
 
-export const CoreLayout = ({ children }) =>
+export const CoreLayout = ({ children }) => (
   <div className={classes.container}>
     <Navbar />
-    <div className={classes.children}>
-      {children}
-    </div>
+    <div className={classes.children}>{children}</div>
   </div>
+)
 
 CoreLayout.propTypes = {
-  children: React.PropTypes.element.isRequired
+  children: PropTypes.element.isRequired
 }
 
 export default CoreLayout
