@@ -1,51 +1,55 @@
-import React, { Component, cloneElement, PropTypes } from 'react'
-import { map } from 'lodash'
+import React, { Component } from 'react'
+import PropTypes from 'prop-types'
+import { map, get } from 'lodash'
 import { connect } from 'react-redux'
 import {
   firebaseConnect,
   populatedDataToJS,
   pathToJS,
+  dataToJS,
   isLoaded,
   isEmpty
 } from 'react-redux-firebase'
 import { LIST_PATH } from 'constants'
-import { UserIsAuthenticated } from 'utils/router'
 import LoadingSpinner from 'components/LoadingSpinner'
-import ProjectTile from '../components/ProjectTile/ProjectTile'
-import NewProjectTile from '../components/NewProjectTile/NewProjectTile'
-import NewProjectDialog from '../components/NewProjectDialog/NewProjectDialog'
+import ProjectTile from '../components/ProjectTile'
+import NewProjectTile from '../components/NewProjectTile'
+import NewProjectDialog from '../components/NewProjectDialog'
 import classes from './ProjectsContainer.scss'
 
 const populates = [{ child: 'createdBy', root: 'users', keyProp: 'uid' }]
 
-@UserIsAuthenticated
-@firebaseConnect([
-  { path: 'projects', populates }
-  // 'projects#populate=owner:users' // string equivalent
+@firebaseConnect(({ params, auth }) => [
+  {
+    path: 'projects',
+    populates
+  }
 ])
 @connect(({ firebase }, { params }) => ({
-  auth: pathToJS(firebase, 'auth'),
-  projects: populatedDataToJS(firebase, 'projects', populates)
+  projects: populatedDataToJS(firebase, 'projects', populates),
+  unpopulatedProjects: dataToJS(firebase, 'projects'),
+  auth: pathToJS(firebase, 'auth')
 }))
 export default class Projects extends Component {
   static contextTypes = {
-    router: React.PropTypes.object.isRequired
+    router: PropTypes.object.isRequired
   }
 
   static propTypes = {
+    children: PropTypes.object,
+    firebase: PropTypes.object.isRequired,
     projects: PropTypes.object,
-    firebase: PropTypes.object,
-    auth: PropTypes.object,
-    children: PropTypes.object
+    unpopulatedProjects: PropTypes.object,
+    auth: PropTypes.object
   }
 
   state = {
-    newProjectModal: false,
-    addProjectModal: false
+    newProjectModal: false
   }
 
   newSubmit = newProject => {
     const { firebase: { pushWithMeta } } = this.props
+    // push new project with createdBy and createdAt
     return pushWithMeta('projects', newProject)
       .then(() => this.setState({ newProjectModal: false }))
       .catch(err => {
@@ -54,18 +58,23 @@ export default class Projects extends Component {
       })
   }
 
-  deleteProject = key => {
-    return this.props.firebase.remove(`projects/${key}`).then(() => {
-      // TODO: Show snackbar
-    })
-  }
+  deleteProject = key => this.props.firebase.remove(`projects/${key}`)
 
   toggleModal = (name, project) => {
     this.setState({ [`${name}Modal`]: !this.state[`${name}Modal`] })
   }
 
+  getDeleteVisible = key => {
+    const { auth, unpopulatedProjects } = this.props
+    return (
+      !isEmpty(this.props.auth) &&
+      get(unpopulatedProjects, `${key}.createdBy`) === auth.uid
+    )
+  }
+
   render() {
     const { projects, auth } = this.props
+    const { newProjectModal } = this.state
 
     if (!isLoaded(projects, auth)) {
       return <LoadingSpinner />
@@ -74,36 +83,31 @@ export default class Projects extends Component {
     // Project Route is being loaded
     if (this.props.children) {
       // pass all props to children routes
-      return cloneElement(this.props.children, this.props)
+      return React.cloneElement(this.props.children, this.props)
     }
-
-    const { newProjectModal } = this.state
 
     return (
       <div className={classes.container}>
-        {newProjectModal &&
+        {newProjectModal && (
           <NewProjectDialog
             open={newProjectModal}
             onSubmit={this.newSubmit}
             onRequestClose={() => this.toggleModal('newProject')}
-          />}
+          />
+        )}
         <div className={classes.tiles}>
           <NewProjectTile onClick={() => this.toggleModal('newProject')} />
           {!isEmpty(projects) &&
-            map(projects, (project, key) =>
+            map(projects, (project, key) => (
               <ProjectTile
                 key={`Project-${key}`}
                 project={project}
                 onCollabClick={this.collabClick}
                 onSelect={() => this.context.router.push(`${LIST_PATH}/${key}`)}
                 onDelete={() => this.deleteProject(key)}
-                showDelete={
-                  auth &&
-                  project.createdBy &&
-                  project.createdBy.uid === auth.uid
-                }
+                showDelete={this.getDeleteVisible(key)}
               />
-            )}
+            ))}
         </div>
       </div>
     )
