@@ -4,7 +4,7 @@ const webpackConfig = require('../config/webpack.config')
 const project = require('../config/project.config')
 
 // Wrapper around webpack to promisify its compiler and supply friendly logging
-const webpackCompiler = (webpackConfig) =>
+const webpackCompiler = webpackConfig =>
   new Promise((resolve, reject) => {
     const compiler = webpack(webpackConfig)
 
@@ -34,22 +34,24 @@ const webpackCompiler = (webpackConfig) =>
 
 const compile = () => {
   debug('Starting compiler.')
-  return Promise.resolve()
-    .then(() => webpackCompiler(webpackConfig))
-    // .then(stats => {
-    //   if (stats.warnings.length && project.compiler_fail_on_warning) {
-    //     throw new Error('Config set to fail on warning, exiting with status code "1".')
-    //   }
-    //   debug('Copying static assets to dist folder.')
-    //   fs.copySync(project.paths.public(), project.paths.dist())
-    // })
-    .then(() => {
-      debug('Compilation completed successfully.')
-    })
-    .catch((err) => {
-      debug('Compiler encountered an error.', err)
-      process.exit(1)
-    })
+  return (
+    Promise.resolve()
+      .then(() => webpackCompiler(webpackConfig))
+      // .then(stats => {
+      //   if (stats.warnings.length && project.compiler_fail_on_warning) {
+      //     throw new Error('Config set to fail on warning, exiting with status code "1".')
+      //   }
+      //   debug('Copying static assets to dist folder.')
+      //   fs.copySync(project.paths.public(), project.paths.dist())
+      // })
+      .then(() => {
+        debug('Compilation completed successfully.')
+      })
+      .catch(err => {
+        debug('Compiler encountered an error.', err)
+        process.exit(1)
+      })
+  )
 }
 
 compile()
