@@ -8,8 +8,8 @@ module.exports = {
   // NOTE: In development, we use an explicit public path when the assets
   // are served webpack by to fix this issue:
   // http://stackoverflow.com/questions/34133808/webpack-ots-parsing-error-loading-fonts/34133809#34133809
-  development : (config) => ({
-    compiler_public_path : `http://${config.server_host}:${config.server_port}/`,
+  development: config => ({
+    compiler_public_path: `http://${config.server_host}:${config.server_port}/`,
     firebase: {
       apiKey: 'AIzaSyCTUERDM-Pchn_UDTsfhVPiwM4TtNIxots',
       authDomain: 'redux-firebasev3.firebaseapp.com',
@@ -27,15 +27,15 @@ module.exports = {
   // ======================================================
   // Overrides when NODE_ENV === 'production'
   // ======================================================
-  production : (config) => ({
+  production: config => ({
     compiler_public_path: '/',
     compiler_fail_on_warning: false,
     compiler_hash_type: 'chunkhash',
     compiler_devtool: null,
     compiler_stats: {
-      chunks       : true,
-      chunkModules : true,
-      colors       : true
+      chunks: true,
+      chunkModules: true,
+      colors: true
     },
     firebase: {
       apiKey: 'AIzaSyCTUERDM-Pchn_UDTsfhVPiwM4TtNIxots',
