@@ -9,42 +9,42 @@ debug('Creating default configuration.')
 // Default Configuration
 // ========================================================
 const config = {
-  env : process.env.NODE_ENV || 'development',
+  env: process.env.NODE_ENV || 'development',
 
   // ----------------------------------
   // Project Structure
   // ----------------------------------
-  path_base  : path.resolve(__dirname, '..'),
-  dir_client : 'src',
-  dir_dist   : 'dist',
-  dir_server : 'server',
-  dir_test   : 'tests',
+  path_base: path.resolve(__dirname, '..'),
+  dir_client: 'src',
+  dir_dist: 'dist',
+  dir_server: 'server',
+  dir_test: 'tests',
 
   // ----------------------------------
   // Server Configuration
   // ----------------------------------
-  server_host : ip.address(), // use string 'localhost' to prevent exposure on local network
-  server_port : process.env.PORT || 3000,
+  server_host: ip.address(), // use string 'localhost' to prevent exposure on local network
+  server_port: process.env.PORT || 3000,
 
   // ----------------------------------
   // Compiler Configuration
   // ----------------------------------
-  compiler_babel : {
-    cacheDirectory : true,
-    plugins        : ['transform-runtime', 'lodash', 'transform-decorators-legacy'],
-    presets        : ['es2015', 'react', 'stage-0']
+  compiler_babel: {
+    cacheDirectory: true,
+    plugins: ['transform-runtime', 'lodash', 'transform-decorators-legacy'],
+    presets: ['es2015', 'react', 'stage-0']
   },
-  compiler_devtool         : 'eval',
-  compiler_hash_type       : 'hash',
-  compiler_fail_on_warning : false,
-  compiler_quiet           : false,
-  compiler_public_path     : '/',
-  compiler_stats           : {
-    chunks : false,
-    chunkModules : false,
-    colors : true
+  compiler_devtool: 'eval',
+  compiler_hash_type: 'hash',
+  compiler_fail_on_warning: false,
+  compiler_quiet: false,
+  compiler_public_path: '/',
+  compiler_stats: {
+    chunks: false,
+    chunkModules: false,
+    colors: true
   },
-  compiler_vendors : [
+  compiler_vendors: [
     'react',
     'react-router',
     'react-redux',
@@ -58,9 +58,9 @@ const config = {
   // ----------------------------------
   // Test Configuration
   // ----------------------------------
-  coverage_reporters : [
-    { type : 'text-summary' },
-    { type : 'lcov', dir : 'coverage' }
+  coverage_reporters: [
+    { type: 'text-summary' },
+    { type: 'lcov', dir: 'coverage' }
   ]
 }
 
@@ -78,15 +78,15 @@ Edit at Your Own Risk
 // ------------------------------------
 // N.B.: globals added here must _also_ be added to .eslintrc
 config.globals = {
-  'process.env'  : {
-    'NODE_ENV' : JSON.stringify(config.env)
+  'process.env': {
+    NODE_ENV: JSON.stringify(config.env)
   },
-  'NODE_ENV'     : config.env,
-  '__DEV__'      : config.env === 'development',
-  '__PROD__'     : config.env === 'production',
-  '__TEST__'     : config.env === 'test',
-  '__COVERAGE__' : !argv.watch && config.env === 'test',
-  '__BASENAME__' : JSON.stringify(process.env.BASENAME || '')
+  NODE_ENV: config.env,
+  __DEV__: config.env === 'development',
+  __PROD__: config.env === 'production',
+  __TEST__: config.env === 'test',
+  __COVERAGE__: !argv.watch && config.env === 'test',
+  __BASENAME__: JSON.stringify(process.env.BASENAME || '')
 }
 
 // ------------------------------------
@@ -94,29 +94,28 @@ config.globals = {
 // ------------------------------------
 const pkg = require('../package.json')
 
-config.compiler_vendors = config.compiler_vendors
-  .filter((dep) => {
-    if (pkg.dependencies[dep]) return true
+config.compiler_vendors = config.compiler_vendors.filter(dep => {
+  if (pkg.dependencies[dep]) return true
 
-    debug(
-      `Package "${dep}" was not found as an npm dependency in package.json; ` +
+  debug(
+    `Package "${dep}" was not found as an npm dependency in package.json; ` +
       `it won't be included in the webpack vendor bundle.
        Consider removing it from \`compiler_vendors\` in ~/config/index.js`
-    )
-  })
+  )
+})
 
 // ------------------------------------
 // Utilities
 // ------------------------------------
-function base () {
+function base() {
   const args = [config.path_base].concat([].slice.call(arguments))
   return path.resolve.apply(path, args)
 }
 
 config.paths = {
-  base   : base,
-  client : base.bind(null, config.dir_client),
-  dist   : base.bind(null, config.dir_dist)
+  base: base,
+  client: base.bind(null, config.dir_client),
+  dist: base.bind(null, config.dir_dist)
 }
 
 // ========================================================
