@@ -1,9 +1,10 @@
-import React, { Component, PropTypes } from 'react'
+import React, { Component } from 'react'
+import PropTypes from 'prop-types'
 import { browserHistory, Router } from 'react-router'
 import { Provider } from 'react-redux'
 
 // Themeing/Styling
-import Theme from '../../theme'
+import Theme from 'theme'
 import getMuiTheme from 'material-ui/styles/getMuiTheme'
 
 // Tap Plugin
@@ -15,24 +16,20 @@ export default class AppContainer extends Component {
     muiTheme: PropTypes.object
   }
 
-  getChildContext = () => ({
-    muiTheme: getMuiTheme(Theme)
-  })
-
   static propTypes = {
     routes: PropTypes.object.isRequired,
     store: PropTypes.object.isRequired
   }
 
+  getChildContext = () => ({
+    muiTheme: getMuiTheme(Theme)
+  })
+
   render() {
     const { routes, store } = this.props
     return (
       <Provider store={store}>
-        <div style={{ height: '100%' }}>
-          <Router history={browserHistory}>
-            {routes}
-          </Router>
-        </div>
+        <Router history={browserHistory}>{routes}</Router>
       </Provider>
     )
   }
