@@ -1,6 +1,13 @@
-import React, { Component, PropTypes } from 'react'
-import classes from './Navbar.scss'
+import React, { Component } from 'react'
+import PropTypes from 'prop-types'
 import { Link } from 'react-router'
+import AppBar from 'material-ui/AppBar'
+import IconMenu from 'material-ui/IconMenu'
+import IconButton from 'material-ui/IconButton'
+import MenuItem from 'material-ui/MenuItem'
+import FlatButton from 'material-ui/FlatButton'
+import DownArrow from 'material-ui/svg-icons/hardware/keyboard-arrow-down'
+import Avatar from 'material-ui/Avatar'
 import { connect } from 'react-redux'
 import {
   firebaseConnect,
@@ -9,16 +16,8 @@ import {
   isEmpty
 } from 'react-redux-firebase'
 import { LIST_PATH, ACCOUNT_PATH, LOGIN_PATH, SIGNUP_PATH } from 'constants'
-
-// Components
-import AppBar from 'material-ui/AppBar'
-import IconMenu from 'material-ui/IconMenu'
-import IconButton from 'material-ui/IconButton'
-import MenuItem from 'material-ui/MenuItem'
-import FlatButton from 'material-ui/FlatButton'
-import DownArrow from 'material-ui/svg-icons/hardware/keyboard-arrow-down'
-import Avatar from 'material-ui/Avatar'
 import defaultUserImage from 'static/User.png'
+import classes from './Navbar.scss'
 
 const buttonStyle = {
   color: 'white',
@@ -39,7 +38,6 @@ const avatarStyles = {
 
 @firebaseConnect()
 @connect(({ firebase }) => ({
-  authError: pathToJS(firebase, 'authError'),
   auth: pathToJS(firebase, 'auth'),
   account: pathToJS(firebase, 'profile')
 }))
@@ -50,6 +48,7 @@ export default class Navbar extends Component {
 
   static propTypes = {
     account: PropTypes.object,
+    auth: PropTypes.object,
     firebase: PropTypes.object.isRequired
   }
 
@@ -59,8 +58,8 @@ export default class Navbar extends Component {
   }
 
   render() {
-    const { account } = this.props
-    const accountExists = isLoaded(account) && !isEmpty(account)
+    const { account, auth } = this.props
+    const authExists = isLoaded(auth) && !isEmpty(auth)
 
     const iconButton = (
       <IconButton style={avatarStyles.button} disableTouchRipple>
@@ -68,17 +67,17 @@ export default class Navbar extends Component {
           <div className="hidden-mobile">
             <Avatar
               src={
-                accountExists && account.avatarUrl
-                  ? account.avatarUrl
-                  : defaultUserImage
+                account && account.avatarUrl ? (
+                  account.avatarUrl
+                ) : (
+                  defaultUserImage
+                )
               }
             />
           </div>
           <div className={classes['avatar-text']}>
             <span className={`${classes['avatar-text-name']} hidden-mobile`}>
-              {accountExists && account.displayName
-                ? account.displayName
-                : 'User'}
+              {account && account.displayName ? account.displayName : 'User'}
             </span>
             <DownArrow color="white" />
           </div>
@@ -97,34 +96,32 @@ export default class Navbar extends Component {
       </div>
     )
 
-    const rightMenu = accountExists
-      ? <IconMenu
-          iconButtonElement={iconButton}
-          targetOrigin={{ horizontal: 'right', vertical: 'bottom' }}
-          anchorOrigin={{ horizontal: 'right', vertical: 'top' }}
-          animated={false}
-        >
-          <MenuItem
-            primaryText="Account"
-            onTouchTap={() => this.context.router.push(ACCOUNT_PATH)}
-          />
-          <MenuItem primaryText="Sign out" onTouchTap={this.handleLogout} />
-        </IconMenu>
-      : mainMenu
+    const rightMenu = authExists ? (
+      <IconMenu
+        iconButtonElement={iconButton}
+        targetOrigin={{ horizontal: 'right', vertical: 'bottom' }}
+        anchorOrigin={{ horizontal: 'right', vertical: 'top' }}
+        animated={false}>
+        <MenuItem
+          primaryText="Account"
+          onTouchTap={() => this.context.router.push(ACCOUNT_PATH)}
+        />
+        <MenuItem primaryText="Sign out" onTouchTap={this.handleLogout} />
+      </IconMenu>
+    ) : (
+      mainMenu
+    )
 
     return (
       <AppBar
         title={
-          <Link
-            to={accountExists ? `${LIST_PATH}` : '/'}
-            className={classes.brand}
-          >
-            material example
+          <Link to={authExists ? LIST_PATH : '/'} className={classes.brand}>
+            material
           </Link>
         }
         showMenuIconButton={false}
-        iconElementRight={rightMenu}
-        iconStyleRight={accountExists ? avatarStyles.wrapper : {}}
+        iconElementRight={isLoaded(auth, account) ? rightMenu : null}
+        iconStyleRight={authExists ? avatarStyles.wrapper : {}}
         className={classes.appBar}
       />
     )
