@@ -1,22 +1,33 @@
-import React, { PropTypes } from 'react'
+import React from 'react'
+import PropTypes from 'prop-types'
 import { Field, reduxForm } from 'redux-form'
 import RaisedButton from 'material-ui/RaisedButton'
-import TextField from 'components/TextField'
+import { TextField } from 'redux-form-material-ui'
 import { ACCOUNT_FORM_NAME } from 'constants'
 import ProviderDataForm from '../ProviderDataForm'
 import classes from './AccountForm.scss'
 
-export const AccountForm = ({ account, handleSubmit, submitting }) =>
+export const AccountForm = ({ account, handleSubmit, submitting }) => (
   <form className={classes.container} onSubmit={handleSubmit}>
     <h4>Account</h4>
-    <Field name="displayName" component={TextField} label="Display Name" />
-    <Field name="email" component={TextField} label="Email" />
+    <Field
+      name="displayName"
+      component={TextField}
+      floatingLabelText="Display Name"
+    />
+    <Field name="email" component={TextField} floatingLabelText="Email" />
+    <Field
+      name="avatarUrl"
+      component={TextField}
+      floatingLabelText="Avatar Url"
+    />
     {!!account &&
-      !!account.providerData &&
+    !!account.providerData && (
       <div>
         <h4>Linked Accounts</h4>
         <ProviderDataForm providerData={account.providerData} />
-      </div>}
+      </div>
+    )}
     <RaisedButton
       primary
       label="Save"
@@ -24,6 +35,7 @@ export const AccountForm = ({ account, handleSubmit, submitting }) =>
       className={classes.submit}
     />
   </form>
+)
 
 AccountForm.propTypes = {
   account: PropTypes.object,
