@@ -1,31 +1,34 @@
-import React, { PropTypes } from 'react'
+import React from 'react'
+import PropTypes from 'prop-types'
 import { List, ListItem } from 'material-ui/List'
 import classes from './ProviderDataForm.scss'
 import AccountCircle from 'material-ui/svg-icons/action/account-circle'
 
-export const ProviderData = ({ providerData }) =>
+export const ProviderData = ({ providerData }) => (
   <div className={classes.container}>
     <List>
-      {providerData.map((providerAccount, i) =>
+      {providerData.map((providerAccount, i) => (
         <ListItem
           key={i}
           primaryText={providerAccount.providerId}
           leftIcon={<AccountCircle />}
           nestedItems={[
             <ListItem
-              key={`Name-${i}`}
+              key="displayName"
               primaryText={providerAccount.displayName}
             />,
             <ListItem
+              key="email"
               label="email"
-              key={`Email-${i}`}
               primaryText={providerAccount.email}
+              disabled
             />
           ]}
         />
-      )}
+      ))}
     </List>
   </div>
+)
 
 ProviderData.propTypes = {
   providerData: PropTypes.array.isRequired
