@@ -1,54 +1,35 @@
-import React, { Component, PropTypes } from 'react'
+import React, { Component } from 'react'
+import PropTypes from 'prop-types'
 import Paper from 'material-ui/Paper'
 import { connect } from 'react-redux'
 import { firebaseConnect, pathToJS, isLoaded } from 'react-redux-firebase'
-import { reduxFirebase as rfConfig } from 'config'
+import { reduxFirebase as fbReduxSettings } from 'config'
 import { UserIsAuthenticated } from 'utils/router'
 import defaultUserImageUrl from 'static/User.png'
 import LoadingSpinner from 'components/LoadingSpinner'
 import AccountForm from '../components/AccountForm/AccountForm'
 import classes from './AccountContainer.scss'
 
-@UserIsAuthenticated
+@UserIsAuthenticated // redirect to /login if user is not authenticated
 @firebaseConnect()
-@connect(
-  // Map redux state to props
-  ({ firebase }) => ({
-    auth: pathToJS(firebase, 'auth'),
-    account: pathToJS(firebase, 'profile')
-  })
-)
-export default // redirect to /login if user is not authenticated
-// add this.props.firebase
-class Account extends Component {
+@connect(({ firebase }) => ({
+  auth: pathToJS(firebase, 'auth'),
+  account: pathToJS(firebase, 'profile')
+}))
+export default class Account extends Component {
   static propTypes = {
     account: PropTypes.object,
-    auth: PropTypes.shape({
-      uid: PropTypes.string
-    }),
+    auth: PropTypes.object,
     firebase: PropTypes.shape({
-      update: PropTypes.func.isRequired,
       logout: PropTypes.func.isRequired
     })
   }
 
-  state = { modalOpen: false }
-
-  handleLogout = () => this.props.firebase.logout()
-
-  toggleModal = () => {
-    this.setState({
-      modalOpen: !this.state.modalOpen
-    })
+  updateAccount = newAccount => {
+    const { firebase: { update }, auth } = this.props
+    // corresponds to /users/${uid}
+    return update(`${fbReduxSettings.userProfile}/${auth.uid}`, newAccount)
   }
-
-  updateAccount = newData =>
-    this.props.firebase
-      .update(`${rfConfig.userProfile}/${this.props.auth.uid}`, newData)
-      .catch(err => {
-        console.error('Error updating account', err) // eslint-disable-line no-console
-        // TODO: Display error to user
-      })
 
   render() {
     const { account } = this.props
@@ -64,15 +45,20 @@ class Account extends Component {
             <div className={classes.avatar}>
               <img
                 className={classes.avatarCurrent}
-                src={(account && account.avatarUrl) || defaultUserImageUrl}
+                src={
+                  account && account.avatarUrl ? (
+                    account.avatarUrl
+                  ) : (
+                    defaultUserImageUrl
+                  )
+                }
                 onClick={this.toggleModal}
               />
             </div>
             <div className={classes.meta}>
               <AccountForm
-                initialValues={account}
-                account={account}
                 onSubmit={this.updateAccount}
+                initialValues={account}
               />
             </div>
           </div>
