@@ -1,4 +1,5 @@
-import React, { Component, PropTypes } from 'react'
+import React, { Component } from 'react'
+import PropTypes from 'prop-types'
 import classes from './TodoItem.scss'
 import { ListItem } from 'material-ui/List'
 import Checkbox from 'material-ui/Checkbox'
@@ -28,15 +29,15 @@ export default class TodoItem extends Component {
           rightIcon={<Delete onClick={() => onDeleteClick(todo._key || id)} />}
           secondaryText={
             <p>
-              <span className="TodoItem-Text">
-                {todo.text}
-              </span>
+              <span className="TodoItem-Text">{todo.text}</span>
               <br />
               <span className="TodoItem-Owner">
                 Owner:{' '}
-                {isObject(todo.owner)
-                  ? todo.owner.displayName || todo.owner.username
-                  : todo.owner || 'No Owner'}
+                {isObject(todo.owner) ? (
+                  todo.owner.displayName || todo.owner.username
+                ) : (
+                  todo.owner || 'No Owner'
+                )}
               </span>
             </p>
           }
