@@ -1,30 +1,36 @@
-import React, { PropTypes } from 'react'
+import React from 'react'
+import PropTypes from 'prop-types'
 import { Link } from 'react-router'
 import { Field, reduxForm } from 'redux-form'
+import { TextField } from 'redux-form-material-ui'
 import RaisedButton from 'material-ui/RaisedButton'
 import Checkbox from 'material-ui/Checkbox'
 import { RECOVER_PATH, LOGIN_FORM_NAME } from 'constants'
-import TextField from 'components/TextField'
-import { required, validateEmail } from 'utils/forms'
+import { required, validateEmail } from 'utils/form'
 import classes from './LoginForm.scss'
 
-export const LoginForm = ({ handleSubmit, submitting }) =>
+export const LoginForm = ({ pristine, submitting, handleSubmit }) => (
   <form className={classes.container} onSubmit={handleSubmit}>
     <Field
       name="email"
       component={TextField}
-      label="Email"
+      floatingLabelText="Email"
       validate={[required, validateEmail]}
     />
     <Field
       name="password"
       component={TextField}
-      label="Password"
+      floatingLabelText="Password"
       type="password"
-      validate={[required]}
+      validate={required}
     />
     <div className={classes.submit}>
-      <RaisedButton label="Login" primary type="submit" disabled={submitting} />
+      <RaisedButton
+        label={submitting ? 'Loading' : 'Login'}
+        primary
+        type="submit"
+        disabled={pristine || submitting}
+      />
     </div>
     <div className={classes.options}>
       <div className={classes.remember}>
@@ -40,10 +46,12 @@ export const LoginForm = ({ handleSubmit, submitting }) =>
       </Link>
     </div>
   </form>
+)
 
 LoginForm.propTypes = {
-  handleSubmit: PropTypes.func,
-  submitting: PropTypes.bool
+  pristine: PropTypes.bool.isRequired, // added by redux-form
+  submitting: PropTypes.bool.isRequired, // added by redux-form
+  handleSubmit: PropTypes.func.isRequired // added by redux-form
 }
 
 export default reduxForm({
