@@ -1,5 +1,4 @@
 import { LIST_PATH as path } from 'constants'
-import Project from './routes/Project'
 
 export default store => ({
   path,
@@ -22,8 +21,14 @@ export default store => ({
       'Projects'
     )
   },
-  childRoutes: [
-    Project // not function for sync route
-    // async routes definitions require function here i.e. Project(store)
-  ]
+  getChildRoutes(partialNextState, cb) {
+    require.ensure([], require => {
+      /*  Webpack - use require callback to define
+          dependencies for bundling   */
+      const Project = require('./routes/Project').default
+
+      /*  Return getComponent   */
+      cb(null, [Project(store)])
+    })
+  }
 })
