@@ -1,13 +1,13 @@
-import React, { Component, PropTypes } from 'react'
+import React, { Component } from 'react'
+import PropTypes from 'prop-types'
 import { firebaseConnect } from 'react-redux-firebase'
 import Snackbar from 'material-ui/Snackbar'
 import Paper from 'material-ui/Paper'
 import RecoverForm from '../components/RecoverForm'
 import EmailForm from '../components/EmailForm'
 
-@firebaseConnect()
-export default // adds this.props.firebase
-class RecoverContainer extends Component {
+@firebaseConnect() // adds this.props.firebase
+export default class RecoverContainer extends Component {
   static propTypes = {
     firebase: PropTypes.object
   }
